fix: circular dependency
Move DocumentService implementation into core
